backend/src/main/resources/ai/requirements.txt
dnspython==2.7.0
pymongo==4.12.0
python-dotenv==1.1.0
